vista de vendedor agregada

+++ src/app/firebase/firestore.objects.service.ts
@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core';
-import { Firestore, collection, doc, setDoc, collectionData, addDoc, deleteDoc } from '@angular/fire/firestore';
+import { Firestore, collection, doc, addDoc, collectionData, deleteDoc } from '@angular/fire/firestore';
 import { Observable } from 'rxjs';
 
 @Injectable({
@@ -8,25 +8,25 @@ import { Observable } from 'rxjs';
 export class FirestoreObjectsService {
   constructor(private firestore: Firestore) {}
 
-  // Método para obtener la referencia de la colección de listas de un usuario
   private getUserListsCollection(uid: string) {
     return collection(this.firestore, `Users/${uid}/lists`);
   }
 
-  // Crear una nueva lista
+  // Crear nueva lista
   addList(uid: string, list: { name: string; createdAt: Date }): Promise<any> {
-    const userListsCollection = this.getUserListsCollection(uid);
-    return addDoc(userListsCollection, list);
+    if (!uid) throw new Error('UID no definido.');
+    return addDoc(this.getUserListsCollection(uid), list);
   }
 
-  // Obtener todas las listas del usuario
+  // Obtener listas del usuario
   getLists(uid: string): Observable<any[]> {
-    const userListsCollection = this.getUserListsCollection(uid);
-    return collectionData(userListsCollection, { idField: 'id' });
+    if (!uid) throw new Error('UID no definido.');
+    return collectionData(this.getUserListsCollection(uid), { idField: 'id' });
   }
 
-  // Eliminar una lista
+  // Eliminar lista
   deleteList(uid: string, listId: string): Promise<void> {
+    if (!uid || !listId) throw new Error('UID o List ID no definido.');
     const listDoc = doc(this.firestore, `Users/${uid}/lists/${listId}`);
     return deleteDoc(listDoc);
   }
